ADD SEARCHBAR FUNCTIONALITY BASED ON USER INPUT
Add (temporarily) searchBar handler function in productController to handle search requests from the header searchBar.

Bugs: There is an error with the current logic as our product structure is based on brand, model and version codes as opposed to names. This will require further logic to be added to match the search text with brand, model and version values.